requirements.txt
emoji==2.14.1
httpx==0.28.1
huggingface-hub==0.31.1
joblib==1.5.0
lightgbm==4.6.0
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
spacy==3.8.5
symspellpy==6.9.0
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
wordcloud==1.9.4
